requirements.txt
Django
djangorestframework
drf-yasg
psycopg2-binary
